backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lejonet_backend"]

[project]
name = "lejonet-backend"
version = "0.1.0"
description = "Lejonet HTR backend — catalog search and transcription storage"
requires-python = ">=3.11"
license = "Apache-2.0"
dependencies = [
    "fastapi[standard]>=0.115,<1",
    "uvicorn[standard]>=0.30,<1",
    "lancedb>=0.20,<1",
    "pyarrow>=18,<19",
    "huggingface-hub>=0.27,<1",
    "tantivy>=0.22",
    "starlette-exporter>=0.21",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "ruff>=0.14",
]
embeddings = [
    "sentence-transformers>=3,<4",
]
telemetry = [
    "opentelemetry-sdk>=1.20",
    "opentelemetry-exporter-otlp>=1.20",
]

[tool.ruff]
line-length = 160

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "S", "B", "A", "C4", "SIM"]
ignore = ["S101", "S603", "S607"]

[tool.ruff.lint.mccabe]
max-complexity = 15
